writing_level/requirements.txt
absl-py==0.1.12
alabaster==0.7.10
asn1crypto==0.23.0
astor==0.6.2
astroid==1.6.2
Babel==2.5.3
Beaker==1.5.4
beautifulsoup4==4.6.0
bleach==1.5.0
blivet==2.1.11
blivet-gui==2.1.7
Brlapi==0.6.6
certifi==2018.1.18
cffi==1.10.0
chardet==3.0.4
cloudpickle==0.5.2
coverage==4.4.2
cryptography==2.0.2
cssselect==0.9.2
cupshelpers==1.0
cycler==0.10.0
decorator==4.2.1
docutils==0.14
entrypoints==0.2.3
enum34==1.1.6
fros==1.1
gast==0.2.0
gpg==1.10.0
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
humanize==0.5.1
idna==2.6
imagesize==1.0.0
iniparse==0.4
IPy==0.81
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
isc==2.0
isort==4.3.4
javapackages==5.0.0.dev1
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
Keras==2.1.5
langtable==0.0.38
lazy-object-proxy==1.3.1
lxml==4.1.1
Mako==1.0.6.dev0
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.0.0
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.2.5
ntplib==0.3.3
numpy==1.14.2
numpydoc==0.7.0
olefile==0.45.1
ordered-set==2.0.0
packaging==17.1
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
pid==2.1.1
Pillow==4.3.0
ply==3.9
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
psutil==5.4.3
ptyprocess==0.5.2
pwquality==1.4.0
pycairo==1.15.3
pycodestyle==2.3.1
pycparser==2.14
pycups==1.9.72
pycurl==7.43.0
pydbus==0.6.0
pyenchant==1.6.10
pyflakes==1.6.0
Pygments==2.2.0
pygobject==3.26.1
pygpgme==0.3
pyinotify==0.9.6
PyIscsi==1.0
pykickstart==2.41
pylint==1.8.3
PyOpenGL==3.1.0
pyOpenSSL==17.2.0
pyparsing==2.2.0
pyparted==3.11.0
PyQt5==5.9.2
PySocks==1.6.7
python-augeas==0.5.0
python-dateutil==2.7.0
python-dmidecode==3.12.2
python-meh==0.43
pytz==2018.3
pyudev==0.21.0
pyxdg==0.25
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.0
requests==2.18.4
requests-file==1.4
requests-ftp==0.3.1
rope==0.10.7
rope-py3k==0.9.4.post1
rpm==4.14.1
scikit-learn==0.19.1
scipy==1.0.0
sepolicy==1.1
setools==4.1.1
setroubleshoot==1.1
simplegeneric==0.8.1
simpleline==0.6
sip==4.19.8
six==1.11.0
sklearn==0.0
slip==0.6.4
slip.dbus==0.6.4
snowballstemmer==1.2.1
sos==3.5
Sphinx==1.7.2
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
SSSDConfig==1.16.1
style==1.1.0
systemd-python==234
tensorboard==1.6.0
tensorflow==1.6.0
tensorflow-gpu==1.5.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
testpath==0.3.1
tornado==5.0.1
traitlets==4.3.2
update==0.0.1
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.10.11
xmltodict==0.11.0
